requirements.txt
angus-sdk-python==0.0.10
colorama==0.3.7
httplib2==0.9.2
imutils==0.3.6
numpy==1.11.1
opencv-python==3.1.0.0
py==1.4.32
PySocks==1.6.5
pytest==2.6.4
python-dateutil==2.6.0
pytz==2016.10
requests==2.5.3
requests-futures==0.9.7
six==1.10.0
twilio==5.6.0
